Code clean-up

+++ GreatGine/GGApplication.cpp
@@ -6,7 +6,7 @@
 GGApplication::GGApplication( HINSTANCE _hInstance, GGConfig& _config )
 	:
 	m_title( L"GreatGine" ),
-	m_version( L"0.1.1 DEV"),
+	m_version( L"0.1.1 DEV" ),
 	m_running( false ),
 	m_exitKey( _config.GetUint( "key_exit", VK_ESCAPE ) ),
 
@@ -15,7 +15,7 @@ GGApplication::GGApplication( HINSTANCE _hInstance, GGConfig& _config )
 	m_graphics( m_window, _config ),
 	m_game( m_input, _config ),
 	m_world(),
-	m_gui( m_graphics.GetResolutionX(), m_graphics .GetResolutionY() )
+	m_gui( m_graphics.GetResolutionX(), m_graphics.GetResolutionY() )
 {
 	m_input.RegisterHandler( this );
 	m_input.RegisterHandler( &m_graphics );
@@ -50,7 +50,7 @@ void GGApplication::Run()
 void GGApplication::Update()
 {
 	m_timer.Update();
-	float frameTime = m_timer.GetFrameTime();
+	double frameTime = m_timer.GetFrameTime();
 
 	m_game.Update( m_world, frameTime );
 	m_graphics.Update( m_world, frameTime );
