+++ GreatGine/GGFPSCounter.cpp
@@ -1,7 +1,7 @@
 #include "PCH.h"
 #include "GGFPSCounter.h"
 
-void GGFPSCounter::Update( float _frameTime )
+void GGFPSCounter::Update( double _frameTime )
 {
 	const int accumulatorSize = 30;
 	if( m_measurementCount < accumulatorSize )
@@ -13,9 +13,9 @@ void GGFPSCounter::Update( float _frameTime )
 	else
 	{
 		//m_fps = m_fps * 0.1f + ((1.0f * accumulatorSize) / m_frameTimeAccumulator) * 0.9f;
-		m_fps = (1.0f * accumulatorSize) / m_frameTimeAccumulator;
+		m_fps = static_cast<float>((1.0f * accumulatorSize) / m_frameTimeAccumulator);
 		m_measurementCount = 0;
-		m_frameTimeAccumulator = 0.0f;
+		m_frameTimeAccumulator = 0.0;
 	}
 
 	return;
